Restructure FileList. Update edit file name logic.
